+++ clarifai/rag/rag.py
@@ -1,3 +1,4 @@
+import uuid
 from datetime import datetime
 from typing import List
 
@@ -10,9 +11,13 @@ from clarifai.client.input import Inputs
 from clarifai.client.model import Model
 from clarifai.client.user import User
 from clarifai.client.workflow import Workflow
+from clarifai.constants.rag import MAX_UPLOAD_BATCH_SIZE
 from clarifai.errors import UserError
-from clarifai.rag.utils import convert_messages_to_str, format_assistant_message
+from clarifai.rag.utils import (convert_messages_to_str, format_assistant_message, load_documents,
+                                split_document)
 from clarifai.utils.logging import get_logger
+
+DEFAULT_RAG_PROMPT_TEMPLATE = "Context information is below:\n{data.hits}\nGiven the context information and not prior knowledge, answer the query.\nQuery: {data.text.raw}\nAnswer: "
 
 
 class RAG:
@@ -21,7 +26,8 @@ class RAG:
 
     Example:
         >>> from clarifai.rag import RAG
-        >>> rag_agent = RAG()
+        >>> rag_agent = RAG(workflow_url=YOUR_WORKFLOW_URL)
+        >>> rag_agent.chat(messages=[{"role":"human", "content":"What is Clarifai"}])
     """
   chat_state_id = None
 
@@ -46,6 +52,7 @@ class RAG:
   @classmethod
   def setup(cls,
             user_id: str = None,
+            app_url: str = None,
             llm_url: str = "https://clarifai.com/mistralai/completion/models/mistral-7B-Instruct",
             base_workflow: str = "Text",
             workflow_yaml_filename: str = 'prompter_wf.yaml',
@@ -54,24 +61,51 @@ class RAG:
             **kwargs):
     """Creates an app with `Text` as base workflow, create prompt model, create prompt workflow.
 
+    **kwargs: Additional keyword arguments to be passed to rag-promter model.
+          - min_score (float): The minimum score for search hits.
+          - max_results (float): The maximum number of search hits.
+          - prompt_template (str): The prompt template used. Must contain {data.hits} for the search hits and {data.text.raw} for the query string.
+
     Example:
         >>> from clarifai.rag import RAG
-        >>> rag_agent = RAG.setup()
+        >>> rag_agent = RAG.setup(user_id=YOUR_USER_ID)
+        >>> rag_agent.chat(messages=[{"role":"human", "content":"What is Clarifai"}])
+
+    Or if you already have an existing app with ingested data:
+        >>> rag_agent = RAG.setup(app_url=YOUR_APP_URL)
+        >>> rag_agent.chat(messages=[{"role":"human", "content":"What is Clarifai"}])
     """
-    user = User(user_id=user_id, base_url=base_url, pat=pat)
+
+    if user_id and not app_url:
+      user = User(user_id=user_id, base_url=base_url, pat=pat)
+      ## Create an App
+      now_ts = str(int(datetime.now().timestamp()))
+      app_id = f"rag_app_{now_ts}"
+      app = user.create_app(app_id=app_id, base_workflow=base_workflow)
+
+    if not user_id and app_url:
+      app = App(url=app_url, pat=pat)
+
+    if user_id and app_url:
+      raise UserError("Must provide one of user_id or app_url, not both.")
+
+    if not user_id and not app_url:
+      raise UserError(
+          "user_id or app_url must be provided. The user_id can be found at https://clarifai.com/settings."
+      )
+
     llm = Model(llm_url)
 
+    min_score = kwargs.get("min_score", 0.95)
+    max_results = kwargs.get("max_results", 5)
+    prompt_template = kwargs.get("prompt_template", DEFAULT_RAG_PROMPT_TEMPLATE)
     params = Struct()
     params.update({
-        "prompt_template":
-            "Context information is below:\n{data.hits}\nGiven the context information and not prior knowledge, answer the query.\nQuery: {data.text.raw}\nAnswer: "
+        "min_score": min_score,
+        "max_results": max_results,
+        "prompt_template": prompt_template
     })
     prompter_model_params = {"params": params}
-
-    ## Create an App
-    now_ts = str(int(datetime.now().timestamp()))
-    app_id = f"rag_app_{now_ts}"
-    app = user.create_app(app_id=app_id, base_workflow=base_workflow)
 
     ## Create rag-prompter model and version
     prompter_model = app.create_model(
@@ -111,20 +145,113 @@ class RAG:
     del user, llm, prompter_model, prompter_model_params
     return cls(workflow=wf)
 
-  # TODO: Implement this.
-  def upload():
-    """Does the following:
+  def upload(self,
+             file_path: str = None,
+             folder_path: str = None,
+             url: str = None,
+             batch_size: int = 128,
+             chunk_size: int = 1024,
+             chunk_overlap: int = 200,
+             **kwargs) -> None:
+    """Uploads documents to the app.
         - Read from a local directory or public url or local filename.
         - Parse the document(s) into chunks.
         - Ingest chunks into the app with metadata.
 
+    Args:
+        file_path str: File path to the document.
+        folder_path str: Folder path to the documents.
+        url str: Public url to the document.
+        batch_size int: Batch size for uploading.
+        chunk_size int: Chunk size for splitting the document.
+        chunk_overlap int: The token overlap of each chunk when splitting.
+        **kwargs: Additional arguments for the SentenceSplitter. Refer https://docs.llamaindex.ai/en/stable/api/llama_index.node_parser.SentenceSplitter.html
+
     Example:
         >>> from clarifai.rag import RAG
-        >>> rag_agent = RAG().setup()
-        >>> rag_agent.upload("~/work/docs")
-        >>> rag_agent.upload("~/work/docs/manual.pdf")
+        >>> rag_agent = RAG.setup(user_id=YOUR_USER_ID)
+        >>> rag_agent.upload(folder_path = "~/work/docs")
+        >>> rag_agent.upload(file_path = "~/work/docs/manual.pdf")
+        >>> rag_agent.chat(messages=[{"role":"human", "content":"What is Clarifai"}])
     """
-    pass
+    #set batch size
+    if batch_size > MAX_UPLOAD_BATCH_SIZE:
+      raise ValueError(f"batch_size cannot be greater than {MAX_UPLOAD_BATCH_SIZE}")
+
+    #check if only one of file_path, folder_path, or url is specified
+    if file_path and (folder_path or url):
+      raise ValueError("Only one of file_path, folder_path, or url can be specified.")
+    if folder_path and (file_path or url):
+      raise ValueError("Only one of file_path, folder_path, or url can be specified.")
+    if url and (file_path or folder_path):
+      raise ValueError("Only one of file_path, folder_path, or url can be specified.")
+
+    #loading documents
+    documents = load_documents(file_path=file_path, folder_path=folder_path, url=url)
+
+    #splitting documents into chunks
+    text_chunks = []
+    metadata = []
+
+    #iterate through documents
+    for doc in documents:
+      cur_text_chunks = split_document(
+          text=doc.text, chunk_size=chunk_size, chunk_overlap=chunk_overlap, **kwargs)
+      text_chunks.extend(cur_text_chunks)
+      metadata.extend([doc.metadata for _ in range(len(cur_text_chunks))])
+      #if batch size is reached, upload the batch
+      if len(text_chunks) > batch_size:
+        for idx in range(0, len(text_chunks), batch_size):
+          if idx + batch_size > len(text_chunks):
+            continue
+          batch_texts = text_chunks[0:batch_size]
+          batch_ids = [uuid.uuid4().hex for _ in range(batch_size)]
+          #metadata
+          batch_metadatas = metadata[0:batch_size]
+          meta_list = []
+          for meta in batch_metadatas:
+            meta_struct = Struct()
+            meta_struct.update(meta)
+            meta_list.append(meta_struct)
+          del batch_metadatas
+          #creating input proto
+          input_batch = [
+              self._app.inputs().get_text_input(
+                  input_id=batch_ids[i],
+                  raw_text=text,
+                  metadata=meta_list[i],
+              ) for i, text in enumerate(batch_texts)
+          ]
+          #uploading input with metadata
+          self._app.inputs().upload_inputs(inputs=input_batch)
+          #delete uploaded chunks
+          del text_chunks[0:batch_size]
+          del metadata[0:batch_size]
+
+    #uploading the remaining chunks
+    if len(text_chunks) > 0:
+      batch_size = len(text_chunks)
+      batch_ids = [uuid.uuid4().hex for _ in range(batch_size)]
+      #metadata
+      batch_metadatas = metadata[0:batch_size]
+      meta_list = []
+      for meta in batch_metadatas:
+        meta_struct = Struct()
+        meta_struct.update(meta)
+        meta_list.append(meta_struct)
+      del batch_metadatas
+      #creating input proto
+      input_batch = [
+          self._app.inputs().get_text_input(
+              input_id=batch_ids[i],
+              raw_text=text,
+              metadata=meta_list[i],
+          ) for i, text in enumerate(text_chunks)
+      ]
+      #uploading input with metadata
+      self._app.inputs().upload_inputs(inputs=input_batch)
+      del text_chunks
+      del metadata
 
   def chat(self, messages: List[dict], client_manage_state: bool = False) -> List[dict]:
     """Chat interface in OpenAI API format.