+++ clarifai/rag/utils.py
@@ -1,4 +1,11 @@
+import io
+from pathlib import Path
 from typing import List
+
+import requests
+from llama_index import Document, SimpleDirectoryReader, download_loader
+from llama_index.node_parser.text import SentenceSplitter
+from pypdf import PdfReader
 
 
 ## TODO: Make this token-aware.
@@ -26,3 +33,70 @@ def convert_messages_to_str(messages: List[dict]) -> str:
 
 def format_assistant_message(raw_text: str) -> dict:
   return {"role": "assistant", "content": raw_text}
+
+
+def load_documents(file_path: str = None, folder_path: str = None,
+                   url: str = None) -> List[Document]:
+  """Loads documents from a local directory or public url or local filename.
+
+  Args:
+      file_path (str): The path to the filename.
+      folder_path (str): The path to the folder.
+      url (str): The url to the file.
+  """
+  #document loaders for filepath
+  if file_path:
+    if file_path.endswith(".pdf"):
+      PDFReader = download_loader("PDFReader")
+      loader = PDFReader()
+      documents = loader.load_data(file=Path(file_path))
+    elif file_path.endswith(".docx"):
+      docReader = download_loader("DocxReader")
+      loader = docReader()
+      documents = loader.load_data(file=Path(file_path))
+    elif file_path.endswith(".txt"):
+      with open(file_path, 'r') as file:
+        text_content = file.read()
+      documents = [Document(text=text_content)]
+    else:
+      raise ValueError("Only .pdf, .docx, and .txt files are supported.")
+
+  #document loaders for folderpath
+  if folder_path:
+    documents = SimpleDirectoryReader(
+        input_dir=Path(folder_path), required_exts=[".pdf", ".docx"]).load_data()
+
+  #document loaders for url
+  if url:
+    response = requests.get(url)
+    if response.status_code != 200:
+      raise ValueError(f"Invalid url {url}.")
+    #for text files
+    try:
+      documents = [Document(text=response.content)]
+    #for pdf files
+    except Exception:
+      documents = []
+      pdf_file = PdfReader(io.BytesIO(response.content))
+      num_pages = len(pdf_file.pages)
+      for page in range(num_pages):
+        page_text = pdf_file.pages[page].extract_text()
+        documents.append(Document(text=page_text))
+    else:
+      raise ValueError(f"Invalid url {url}.")
+
+  return documents
+
+
+def split_document(text: str, chunk_size: int, chunk_overlap: int, **kwargs) -> List[str]:
+  """Splits a document into chunks of text.
+
+  Args:
+      text (str): The text to split.
+      chunk_size (int): The size of each chunk.
+      chunk_overlap (int): The amount of overlap between each chunk.
+      **kwargs: Additional keyword arguments for the SentenceSplitter.
+  """
+  text_parser = SentenceSplitter(chunk_size=chunk_size, chunk_overlap=chunk_overlap, **kwargs)
+  text_chunks = text_parser.split_text(text)
+  return text_chunks